# internal/app/domain/librarian.go
package domain

import "time"

type Librarian struct {
	Email     string    `json:"email"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type CreateLibrarianPayload struct {
	Email    string `json:"email" binding:"required,email"`
	Name     string `json:"name" binding:"required"`
	Password string `json:"password" binding:"required,min=6"`
}

type CreateUserPayload struct {
	Email string `json:"email" binding:"required,email"`
}